requirements.txt
# Core dependencies for FlightHunter
firebase-admin
python-dotenv
requests
schedule

# API specific
amadeus
